refactor: adding a state prices who will contain the prices of token
