+++ src/msg/instantiate_msg.rs
@@ -1,7 +1,9 @@
 use crate::types::Order;
 use cosmwasm_schema::cw_serde;
+use cosmwasm_std::Coin;
 
 #[cw_serde]
 pub struct InstantiateMsg {
     pub orders: Vec<Order>,
+    pub prices: Vec<Coin>,
 }
