+++ src/tests/cancel_order/successful_cancel_order_with_created_order.rs
@@ -10,6 +10,11 @@ fn successful_cancel_order_with_created_order() {
             .unwrap()
     });
 
+    let instantiate_msg = InstantiateMsg {
+        orders: vec![],
+        prices: vec![],
+    };
+
     let code = ContractWrapper::new(execute, instantiate, query);
     let code_id = app.store_code(Box::new(code));
 
@@ -17,7 +22,7 @@ fn successful_cancel_order_with_created_order() {
         .instantiate_contract(
             code_id,
             Addr::unchecked("owner"),
-            &InstantiateMsg { orders: vec![] },
+            &instantiate_msg,
             &[],
             "Contract",
             None,
