+++ src/tests/cancel_order/unauthorized.rs
@@ -6,6 +6,7 @@ fn unauthorized() {
 
     let instantiate_msg = InstantiateMsg {
         orders: vec![Order::new_dummy()],
+        prices: vec![],
     };
     let id = instantiate_msg.orders[0].order_id.clone().to_owned();
