+++ src/tests/create_order/coin_number.rs
@@ -4,7 +4,11 @@ use super::*;
 fn coin_number() {
     let mut app = App::default();
 
-    let instantiate_msg = InstantiateMsg { orders: vec![] };
+    let instantiate_msg = InstantiateMsg {
+        orders: vec![],
+        prices: vec![],
+    };
+
     let code = ContractWrapper::new(execute, instantiate, query);
     let code_id = app.store_code(Box::new(code));
     let addr = app
