+++ src/tests/get_order/successful_query_message.rs
@@ -7,6 +7,7 @@ fn successful_query_message() {
 
     let instantiate_msg = InstantiateMsg {
         orders: vec![Order::new_dummy()],
+        prices: vec![],
     };
     let id = instantiate_msg.orders[0].order_id.clone().to_owned();
